requirements.txt
click==8.1.7
et_xmlfile==2.0.0
iniconfig==2.0.0
-e git+https://github.com/BrennanTM/nda-pipeline.git@a1112e2317680b5ade3fc8ead29edcd66447d3ad#egg=nda_validator
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tqdm==4.67.1
tzdata==2024.2
